pyproject.toml
[project]
name = "bankfetch"
version = "0.1.0"
description = "CLI for fetching balances and transactions from Enable Banking"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "cryptography>=45.0.0,<46.0.0",
  "httpx>=0.28.0,<0.29.0",
  "pydantic>=2.11.0,<3.0.0",
  "PyJWT>=2.10.0,<3.0.0",
  "PyYAML>=6.0.0,<7.0.0",
  "typer>=0.16.0,<0.17.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "pytest-mock>=3.14.0,<4.0.0",
  "respx>=0.22.0,<0.23.0",
]

[project.scripts]
bankfetch = "bankfetch.cli:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bankfetch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
